Functions revised. Fails in ft_print_u and ft_print_p

+++ ft_printf/ft_print_p.c
@@ -3,19 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_print_p.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gfredes- <gfredes-@student.42malaga.com>   +#+  +:+       +#+        */
+/*   By: gfredes- <gfredes-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/20 22:01:21 by gfredes-          #+#    #+#             */
-/*   Updated: 2023/01/20 22:01:21 by gfredes-         ###   ########.fr       */
+/*   Updated: 2023/01/21 23:56:46 by gfredes-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "ft_printf.h"
 
-#include "libftprinf.h"
-
-void  ft_print_p(unsigned int nb, size_t *len)
+void	ft_print_p(unsigned long int nb, size_t *len)
 {
-  ft_print_c('0', len);
-  ft_print_c('x', len);
-  ft_print_x(nb, len);
+	ft_print_c('0', len);
+	ft_print_c('x', len);
+	ft_print_x(nb, len);
 }
